move ChimeTestClient into its own file
